+++ project/budgets/forms.py
@@ -1,6 +1,7 @@
 from django import forms
 from django.forms import modelformset_factory
 from .models import Budget
+from decimal import Decimal, ROUND_HALF_UP
 
 class BudgetForm(forms.ModelForm):
     class Meta:
@@ -8,8 +9,18 @@ class BudgetForm(forms.ModelForm):
         fields = ['category', 'limit']
         widgets = {
             'category': forms.HiddenInput(),
-            'limit': forms.NumberInput(attrs={'class': 'form-control', 'min': 0, 'step': '0.01'}),
+            'limit': forms.NumberInput(attrs={
+                'class': 'form-control',
+                'min': 0, 
+                'max': 1000000,
+                'oninput': 'validateDecimalPlaces(this)'
+            }),
         }
+
+    def clean_limit(self):
+        limit = self.cleaned_data['limit']
+        rounded_limit = limit.quantize(Decimal('0.01'), rounding=ROUND_HALF_UP)
+        return rounded_limit
 
 BudgetFormSet = modelformset_factory(
     Budget,